requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
biopython==1.85
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==7.1.2
conllu==6.0.0
cymem==2.0.11
en_ner_bc5cdr_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_ner_bc5cdr_md-0.4.0.tar.gz#sha256=694a32a6c3d44ae82f3a6bdf67b3f956a82a4803992a361d537eb6bffa1b549a
en_ner_jnlpba_md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_ner_jnlpba_md-0.4.0.tar.gz#sha256=b58a302c4492e46de7a6ea2660c365c2c5a42d1caf8e3bb2a78b81f2fa126040
exceptiongroup==1.2.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
liftover==1.3.2
lxml==5.3.2
MarkupSafe==3.0.2
murmurhash==1.0.12
narwhals==1.35.0
nmslib==2.1.1
numpy==1.23.5
packaging==24.2
pandas==2.2.3
-e git+https://github.com/carmmmm/paper2kb.git@639bfc2c4c43416ec8ca119bfb5ead6b95740381#egg=paper2kb
pathlib_abc==0.1.1
pathy==0.11.0
pillow==11.2.1
pluggy==1.5.0
preshed==3.0.9
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pybind11==2.6.1
pydantic==1.8.2
pydeck==0.9.1
pyliftover==0.4.1
PyMuPDF==1.25.5
pysbd==0.3.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
scispacy==0.4.0
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
soupsieve==2.6
spacy==3.0.9
spacy-legacy==3.0.12
srsly==2.5.1
streamlit==1.44.1
tenacity==9.1.2
thinc==8.0.17
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.3.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==0.10.1
